Moved registration logic to custom hook and added styling for success and error response
