+++ client/src/components/Register/styles.module.css
@@ -1,0 +1,27 @@
+/* src/styles.css (or your global CSS file) */
+.messageBox {
+    padding: 10px;
+    margin-top: 10px;
+    border-radius: 5px;
+    font-weight: bold;
+  }
+  
+  .error {
+    background-color: #ffe6e6; /* Light red */
+    color: #cc0000; /* Darker red */
+    border: 1px solid #ff9999; /* Slightly darker red border */
+  }
+  
+  .success {
+    background-color: #e6ffe6; /* Light green */
+    color: #009900; /* Darker green */
+    border: 1px solid #99ff99; /* Slightly darker green border */
+  }
+  
+  .pending {
+      background-color: #f0f0f0; /* Light gray */
+      color: #808080; /* Darker gray */
+      border: 1px solid #ccc; /* Slightly darker gray border */
+  }
+  
+  /* Add any other general styles for your registration form, etc. here */
